Avoid code duplication. Move ScoreContributionsTask() into its own file and scoreContributions() into SharedTreeModel
